backend/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
joblib
pydantic
python-multipart
httpx
python-dotenv
